Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
Pillow
numpy
google-generativeai
httpx